+++ switch_model/balancing/renewable_target.py
@@ -27,7 +27,7 @@ def define_components(mod):
     renewable_target[p] is a parameter that describes how much of load (as a percentage) must be served by GENERATION_PROJECTS.
     This assumes that all GENERATION_PROJECTS are renewable generators that count toward the goal.
 
-    system_power_cost[z,t] is a parameter that describes the cost of system/grid power ($/MWh) for each timepoint in each load zone. 
+    hedge_cost[z,t] is a parameter that describes the cost of system/grid power ($/MWh) for each timepoint in each load zone. 
     This could describe wholesale prices of electricity at the demand node or your utility rate structure 
     (whereever you will be buying energy from if not procuring from one of the GENERATION_PROJECTS)
 
@@ -65,11 +65,11 @@ def define_components(mod):
         within=PercentFraction)
 
     # if no system power cost is specified, set the cost to a very small amount
-    # This discourages use of system power in 
-    mod.system_power_cost = Param(
+    # This discourages use of system power 
+    mod.hedge_cost = Param(
         mod.ZONE_TIMEPOINTS,
         within=Reals,
-        default=0.000000001)
+        default=0.0000001)
 
     mod.tp_in_subset = Param(mod.TIMEPOINTS, within=Boolean, default=False)
     
@@ -84,7 +84,7 @@ def define_components(mod):
     #calculate the cost of using system power for the objective function
     mod.SystemPowerCost = Expression(
         mod.TIMEPOINTS,
-        rule=lambda m, t: sum(m.SystemPower[z, t] * m.system_power_cost[z, t] for z in m.LOAD_ZONES) 
+        rule=lambda m, t: sum(m.SystemPower[z, t] * m.hedge_cost[z, t] for z in m.LOAD_ZONES) 
     )
     mod.Cost_Components_Per_TP.append('SystemPowerCost')
     
@@ -149,8 +149,8 @@ def load_inputs(mod, switch_data, inputs_dir):
     renewable_target.csv
         period, renewable_target
 
-    system_power_cost.csv
-        load_zone, timepoint, system_power_cost
+    hedge_cost.csv
+        load_zone, timepoint, hedge_cost
 
     days.csv
         timepoint_id, tp_day, tp_in_subset
@@ -164,10 +164,10 @@ def load_inputs(mod, switch_data, inputs_dir):
 
     #load inputs which include costs for each timepoint in each zone
     switch_data.load_aug(
-        filename=os.path.join(inputs_dir, 'system_power_cost.csv'),
-        select=('load_zone','timepoint','system_power_cost'),
+        filename=os.path.join(inputs_dir, 'hedge_cost.csv'),
+        select=('load_zone','timepoint','hedge_cost'),
         index=mod.ZONE_TIMEPOINTS,
-        param=[mod.system_power_cost])
+        param=[mod.hedge_cost])
 
     # load optional data specifying subset days
     switch_data.load_aug(
@@ -188,11 +188,11 @@ def post_solve(instance, outdir):
         "period": instance.tp_period[t],
         "load_zone": z,
         "System_Power_MW":value(instance.SystemPower[z,t]),
-        "System_Power_Cost_per_MWh":instance.system_power_cost[z,t],
+        "hedge_cost_per_MWh":instance.hedge_cost[z,t],
         "System_Power_GWh_per_year": value(
             instance.SystemPower[z,t] * instance.tp_weight_in_year[t] / 1000),
-        "Annual_System_Power_Cost": value(
-            instance.SystemPower[z,t] * instance.system_power_cost[z,t] *
+        "Annual_hedge_cost": value(
+            instance.SystemPower[z,t] * instance.hedge_cost[z,t] *
             instance.tp_weight_in_year[t])
     } for z, t in instance.ZONE_TIMEPOINTS ]
     SP_df = pd.DataFrame(system_power_dat)